Get the links to files from the CZVV website

From the default URL, (https://vysledky.cermat.cz/statistika/), the function first gets all the project-year combinations you can possibly set using the form fields Projekt and Rok, then proceeds to fetch the HTMLs returned by the server. Finally, it extracts the links to the files and returns a tibble with project, year, data type (either "item" or "aggregated") and the link to the data file you can use to download the data. To download the files, and store them is a handy file structure, use {cermine}'s function download_data() on the result of mine_links().

Usage

mine_links(
  projects = NULL,
  years = NULL,
  data_type = "item",
  force = FALSE,
  base_url = default_base_url()
)

Arguments

projects

Projects to extract the links for. If NULL (the default), all projects available are used. Possible values are Maturitní zkoušky and Jednotná přijímací zkouška (but that may be a subject to change). You'll get the up-to-date options listed if you type in incorrect option(s). To get the current options, you may also call get_exam_events() first and use the project column from the result.

years

Years to extract the links for. If NULL (the default), all years available are used. You'll get the up-to-date options listed if you type in incorrect option(s). To get the current options, you may also call get_exam_events() first and use the year column from the result. Note that if you opt for some impossible combination of project and year, the function will inform you.

data_type

Type of the data to return the links for. item (the default) extracts only item data for every student, aggregated extracts only aggregated data, and both returns both types.

force

Force a refresh of the data, even if cached. Defaults to FALSE.

base_url

Base URL of the CERMAT website. Defaults to https://vysledky.cermat.cz/statistika/.

Details

Note that Rok is a slight misnomer – it actually refers to the exam "event" (usually it comes in year/season format, e.g. "2025 - Jaro" or "2025 - Podzim"), however, for "Jednotné přijímací zkoušky", all terms are presented under single year option. Note further that naming of Rok field options is not consistent.